refactor(collector): split out the shapes, the corpus walk and the DSL graph
Three more concepts out of `Collector`, which goes 1365 -> 1032. Each is a
different kind of cut, and the third one is a smaller cut than its name
suggests — see below.

`Shapes` (205 lines, 0 call sites)

  The thirteen `Data.define`s and the two core-ancestor tables. Declarations,
  not behaviour: two hundred lines of them sat between the class comment and
  `initialize`, so the first thing a reader met was thirteen structs rather
  than the pass they belong to. Included rather than referenced, so `Storage`
  keeps meaning `Storage` in every method that builds one.

`Corpus` (103 lines, 5 call sites)

  The walk from one file to the files it names but does not declare. Split
  along the line between FINDING and KEEPING: which file to open next, in what
  order, and once each, is the walk; where the shapes it hands back are stored
  stays in the collector, because that is its state and nobody else's.

  The derivation is a block, so the walk never learns what a shape collector
  is. It answers `[shapes, lookups]` rather than letting the walk ask the
  shapes what they look for next: `external_lookups` is `protected`, readable
  by one collector from another and by nothing else, and that is worth keeping
  — so the caller reads it and hands it over.

`DslGraph` (149 lines, 6 call sites)

  Where a DSL call ends up: forwards to keepers, keepers to slots, slots to the
  storage method that fills them. Seven shape collections in, and five of them
  are read by nothing else in the file.

  Built late, in `resolve_replays`, because `keeper` reads the resolved
  delegations and a raw one names a constant that may not have resolved yet.

  `dsl_providers` deliberately did NOT come along, though it is the other half
  of "which DSL answers this call". It reads the declarations, extends and
  superclasses — the state the VISITOR writes — where everything else in the
  graph reads method shapes. Two disjoint sets of state in one object is a
  worse seam than the one it removes, and it belongs with the lexical scope
  whenever that gets extracted.

`initialize` still declares 31 ivars: what is left is the collecting itself,
and untangling that is the phase split (collection / absorption / resolution),
which none of these three attempt.

Suite: 1546 examples, 3 failures, all three the worktree-environment ones that
also fail on main from a worktree (they pass in a normal checkout).
